parse output as float to support both int & float

+++ src/mimarsinan/chip_simulation/execute_nevresim.py
@@ -16,7 +16,7 @@ def execute_simulator(simulator_filename, input_count, num_proc = 4):
     for p in pipes:
         line = p.stdout.readline()
         for val in line.split():
-            output_values.append(int(val)) 
+            output_values.append(float(val)) 
         p.wait()
 
     return output_values
